[Shock] - Deprecate RedirectsPlugin and move it into LinksPlugin (Resolves #204)
